move helper functions to Common
