+++ src/Common/Common.jl
@@ -1,8 +1,15 @@
 module Common
 
 import Thermodynamics as TD
+import CloudMicrophysics as CM
+
+const CMP = CM.Parameters
+const CM0 = CM.Microphysics0M
+const CM1 = CM.Microphysics1M
+const CM2 = CM.Microphysics2M
 
 include("equation_types.jl")
+include("helper_functions.jl")
 include("pysdm_functions.jl")
 
 end
