+++ src/K1D_col_sed/K1DColSedModel.jl
@@ -17,7 +17,6 @@ import ..Common as CO
 import ..K1DModel as K1D
 
 include("./ode_utils.jl")
-include("./helper_functions.jl")
 include("./initial_condition.jl")
 include("./tendency.jl")
